// src/app/models/planets.dto.model.ts
export interface PlanetsDto {
    count: number;
    next: string;
    previous: string;
    results: PlanetDto[];
}

export interface PlanetDto {
    name: string;
    rotation_period: string;
    orbital_period: string;
    diameter: string;
    climate: string;
    gravity: string;
    terrain: string;
    surface_water: string;
    population: string;
    residents: string[];
    films: string[];
    created: string;
    edited: string;
    url: string;
}